Add multi cluster selection
