+++ src/commands/delete.ts
@@ -6,6 +6,7 @@ import {
 
 import {flags} from '@oclif/command';
 import {BaseCommand} from '../base/base-command';
+import {cli} from 'cli-ux';
 
 export default class DeleteCommand extends BaseCommand {
   public static description = 'Delete the current project deployment from the slicknode servers.';
@@ -66,7 +67,12 @@ export default class DeleteCommand extends BaseCommand {
 
     // Run delete migration
     try {
-      await this.getClient().fetch(DELETE_PROJECT_MUTATION, {id: env.id});
+      cli.action.start('Deleting project in cluster');
+      const result = await this.getClient().fetch(DELETE_PROJECT_MUTATION, {id: env.id});
+      cli.action.stop();
+      if (result.errors && result.errors.length) {
+        throw new Error(result.errors[0].message);
+      }
     } catch (e) {
       this.error(chalk.red(`Error deleting project: ${e.message}`));
       return;
@@ -91,7 +97,9 @@ export default class DeleteCommand extends BaseCommand {
 
 export const DELETE_PROJECT_MUTATION = `mutation DeleteProjectMutation($id: ID!) {
   deleteProject(input: {id: $id}) {
-    clientMutationId
+    node {
+      id
+    }
   }
 }
 `;
